+++ components/openapi/src/main/scala/pl/touk/nussknacker/openapi/discovery/SwaggerOpenApiDefinitionDiscovery.scala
@@ -28,7 +28,6 @@ object SwaggerOpenApiDefinitionDiscovery
             None,
             None,
             None,
-            None,
           ).toAsyncHttpClientConfig(None).build()
         )
       )